+++ PowerValidatedSolutions.psm1
@@ -7263,10 +7263,10 @@ Function Export-vRLIJsonSpec {
                                             if ($vcfVersion -eq "4.4.0") { $vrliVersion = "8.6.2" }
                                             if ($vcfVersion -eq "4.4.1") { $vrliVersion = "8.6.2" }
                                             if ($vcfVersion -eq "4.5.0") { $vrliVersion = "8.8.2" }
-                                            if ($vcfVersion -eq "4.5.1") { $vrliVersion = "8.12.0" }
-                                            if ($vcfVersion -eq "4.5.2") { $vrliVersion = "8.12.0" }
-                                            if ($vcfVersion -eq "5.0.0") { $vrliVersion = "8.12.0" }
-                                            if ($vcfVersion -eq "5.1.0") { $vrliVersion = "8.12.0" }
+                                            if ($vcfVersion -eq "4.5.1") { $vrliVersion = "8.14.0" }
+                                            if ($vcfVersion -eq "4.5.2") { $vrliVersion = "8.14.0" }
+                                            if ($vcfVersion -eq "5.0.0") { $vrliVersion = "8.14.0" }
+                                            if ($vcfVersion -eq "5.1.0") { $vrliVersion = "8.14.0" }
                                         } else {
                                             $vrliVersion = $customVersion
                                         }
@@ -9485,10 +9485,10 @@ Function Export-vROPsJsonSpec {
                                                     if ($vcfVersion -eq "4.4.0") { $vropsVersion = "8.6.2"}
                                                     if ($vcfVersion -eq "4.4.1") { $vropsVersion = "8.6.2"}
                                                     if ($vcfVersion -eq "4.5.0") { $vropsVersion = "8.6.3"}
-                                                    if ($vcfVersion -eq "4.5.1") { $vropsVersion = "8.12.1"}
-                                                    if ($vcfVersion -eq "4.5.2") { $vropsVersion = "8.12.1"}
-                                                    if ($vcfVersion -eq "5.0.0") { $vropsVersion = "8.12.1"}
-                                                    if ($vcfVersion -eq "5.1.0") { $vropsVersion = "8.12.1"}
+                                                    if ($vcfVersion -eq "4.5.1") { $vropsVersion = "8.14.0"}
+                                                    if ($vcfVersion -eq "4.5.2") { $vropsVersion = "8.14.0"}
+                                                    if ($vcfVersion -eq "5.0.0") { $vropsVersion = "8.14.0"}
+                                                    if ($vcfVersion -eq "5.1.0") { $vropsVersion = "8.14.0"}
                                                 } else {
                                                     $vropsVersion = $customVersion
                                                 }
@@ -11931,10 +11931,10 @@ Function Export-vRAJsonSpec {
                                                     if ($vcfVersion -eq "4.4.0") { $vraVersion = "8.6.2" }
                                                     if ($vcfVersion -eq "4.4.1") { $vraVersion = "8.6.2" }
                                                     if ($vcfVersion -eq "4.5.0") { $vraVersion = "8.8.2" }
-                                                    if ($vcfVersion -eq "4.5.1") { $vraVersion = "8.12.2" }
-                                                    if ($vcfVersion -eq "4.5.2") { $vraVersion = "8.12.2" }
-                                                    if ($vcfVersion -eq "5.0.0") { $vraVersion = "8.12.2" }
-                                                    if ($vcfVersion -eq "5.1.0") { $vraVersion = "8.12.2" }
+                                                    if ($vcfVersion -eq "4.5.1") { $vraVersion = "8.14.0" }
+                                                    if ($vcfVersion -eq "4.5.2") { $vraVersion = "8.14.0" }
+                                                    if ($vcfVersion -eq "5.0.0") { $vraVersion = "8.14.0" }
+                                                    if ($vcfVersion -eq "5.1.0") { $vraVersion = "8.14.0" }
                                                 } else {
                                                     $vraVersion = $customVersion
                                                 }